+++ next.config.mjs
@@ -15,7 +15,7 @@ const withNextIntl = createNextIntlPlugin();
 const nextConfig = {
   webpack: (config) => {
     config.resolve.alias.canvas = false;
-
+    config.resolve.fallback = { fs: false };
     return config;
   },
 };